refactor(construct): LottoService 인터페이스를 사용하도록 하여 구현체와의 의존성 분리
